Add files via upload
The City crawler is kind of broken, but we gave up on that so I haven't iterated further. All the state crawlers so far output every up to date section in each states revised statute that have the key word 'energy'. The Arkansas, Colorado scripts can't identify the final page yet, so when it reaches the end it just indefinitely keeps pulling the links from that last page. In the interest of time I just check if the automated user has reached the end of the statututes, terminate the function, and use revised_statutes = list(set(revised_statutes).